debugged so that coverage executes
